Refactor SocialLogin classes
Separate the concepts of the social login machinery from the storage
of social identity.

The new `SocialIdentity` class will be the output of social login,
and will not be allowed to exist in an invalid state.

Also disallow social registration without an email. Previously we
allowed manually specifying an email in this case, but I think we
should insist that the email is provided by the social platform.