Templatize album (#1617)
The album code was one of the last remaining places with intermixed
HTML and PHP processing. As with everything else, we break it up into
PHP processing files and then template rendering files.

Add an `Smr\Album` class to help consolidate some of the interaction
with the `album` database table. It is not as comprehensive as the
other data model classes (e.g. Player, Account), because I'm not sure
how helpful it is at this point.

Changed the behavior of the album index.php query params, so that
`nick` looks for an exact match, and `search` looks for a substring
match.

Fixed many issues related to not escaping user-generated content that
was being displayed as HTML.