Require trailing comma in function declaration
Enables the following sniff:

SlevomatCodingStandard.Functions.RequireTrailingCommaInDeclaration

Requiring a comma after every argument in a multi-line function call
is helpful because we don't need to modify the last argument should
we decide to add an argument in the future.

This is enabled by default in SlevomatCodingStandard, but is disabled
by DisallowTrailingCommaInDeclaration in psr2r-sniffer 1.5.0. While
this will be removed in the next major version, we can force it here.
See php-fig-rectified/psr2r-sniffer#31.