Page: add getLandingPage method
Sometimes pages need to forward to either CurrentSector or GamePlay
depending on whether or not the user is currently in a game. The new
Page::getLandingPage method provides this logic to avoid repeating it.

This also fixes an unintended behavior where if you are currently in
a game and navigate to the root URL (e.g. www.smrealms.de), then you
would get put on the GamePlay page while still in a game. This was
because index.php redirects to login.php, and if you already have an
active session, you ultimately forward to LoginProcessor, which then
forwarded explicitly to GamePlay (even if your session had a game).
We now use Page::getLandingPage here to fix this.