GameLeaveProcessor: fix infinite loops when game ends
The `skipRedirect` flag was erroneously removed when we converted to
Page classes in 1361b05b3d37. I think I was thinking that since the
processor removes the session game, we wouldn't need to use
`skipRedirect`; however, since we _forward_ to GameLeaveProcessor, we
still have to go through the `do_voodoo` logic with a session game
before we execute the GameLeaveProcessor, and since we were not
skipping redirects, it would forward to GameLeaveProcessor again in an
infinite loop.

This is now fixed by hard-coding `skipRedirect` in GameLeaveProcessor.
It should be fine to use it for the case that isn't the game ending
(i.e. when the user manually clicks to leave the game).