Add more database column uniqueness constraints
Any column used as a lookup should be unique, otherwise multiple rows
could match a query, leading to ambiguity at best and impersonation at
worst.

The following are now unique:

* account.discord_id
* account.irc_nick
* account.old_account_id
* account.old_account_id2
* alliance.(discord_channel,game_id)

Note that NULL values are allowed to be duplicated in a unique column,
so we use NULL as the default `old_account_id{,2}`. In the live server,
there were also cases of empty string `discord_channel`, so we convert
these to NULL as well.