EditSectorProcessor: fix adding linked locations
The function `addLocationToSector` was only defined in SaveProcessor,
so we were getting undefined function errors when adding linked
locations in EditSectorProcessor. To fix this, we make the function a
method of SmrSector instead of a free function.

We simplify the logic somewhat by making Sector::addLocation return
early if the sector already has the location, thus we don't have to
perform the checks manually each time we use it.

We also no longer limit the logic to HQ locations. If for some reason
we add linked locations to other kinds of locations, this logic will
now work properly.

There was also a bug adding Fed Beacons in the proper location if the
DEFAULT_FED_RADIUS >= 2. This was due to using `$sector` instead of
`$fedSector` when getting the linked sector.