Ship balance changes for game 150
Devastator:
    -25 shields (300 -> 250)
    +25 armour (1800 -> 1850)

The motivation for this change is that the Devastator effectively had
extra shields in certain combat scenarios (+75 against ports and +200
against planets due to the discrete turret damage values). As a ship
with the 2nd highest raw defenses and almost no relevant downsides,
having this additional numerical quirk was an unneeded advantage.

My goal is to foster more variety in the races that players select, and
this should be a sufficiently subtle change that Creonti is still a
competitive race.