Fix improper GROUP BY usage with strict SQL
> Expression of SELECT list is not in GROUP BY clause and contains
> nonaggregated column 't2.author_account_id' which is not functionally
> dependent on columns in GROUP BY clause; this is incompatible with
> sql_mode=only_full_group_by

While `author_account_id` is a non-aggregated column, it should only
have a single value for each `thread_id` due to the use of FIRST_VALUE.
Adding it as a GROUP BY field fixes the error and should not change
the behavior.

I'm not quite as sure about the IpViewResults queries, but it doesn't
really matter how they behave at this point, since this multi checking
feature is almost entirely unused.