Remove exception for negative HOF stats
This exception was added in 69603853ef0 instead of coercing the amount
to a non-negative number. Now we allow negative stats, though this is
only relevant for (Trade:Money:Profit), the only stat that currently
uses `decreaseHOF`. Note that this scenario is triggered for every
player when they make their first purchase of goods (since they have
no positive profit yet).

In MySQL 8.0.17, unsigned float types are deprecated, so we needed to
remove this column attribute soon anyway.