Account: user-friendly createAccount errors
With the `email` and `hof_name` columns unique in the `account` table
as of 7e27198cccd5cc8e7, we should provide better error messages when
those constraints will be violated. We can do this by checking to see
if an account already exists with any of those properties, and throw
a UserError if it does (rather than an "Unexpected Error" when the
query fails).

Also add table locking to protect against race conditions.