Reserve HoF names for player in-game names
When selecting a player name, you will now get an error if it matches
an existing HoF name. This essentially allows players to reserve a
name that they may want to use, and helps prevent spoofing.

If a player uses this feature to hold someone else's name hostage,
then the plan is for players to be able to ask admins to "reclaim"
the name for them.

The "Join Game" page will now auto-populate the player name input with
the HoF name.