Add config to promote libxml warnings to errors
When libxml handles its own errors, it writes a warning that provides
no context about the offending string. We now handle the libxml errors
manually and output the literal line causing the error.

We use the `ENABLE_LIBXML_ERRORS` config option to specify whether the
errors should be fatal or logged as a warning.