Refactor Changelog and switch to BBCode
The changelog was a common source of libxml errors, because it takes
user input, but did not do any escaping (because it often needed
special HTML formatting). Now we will use BBCode instead of HTML, so
that we can safely escape the entire message.

Note that I will have to manually go in and convert all the old HTML
messages to BBCode on the live server.

The `changelog` database querying has been moved into a separate class
called `Changelog` to avoid duplicating the logic (and so we've had to
rename the admin `Changelog` class to `ChangelogAdd`).