bootstrap.php: fix escaping of exception message
When ENABLE_DEBUG is true, we print the exception message on the page
using `<pre>` tags. However, these don't do any escaping, so we need to
manually escape the content in case it includes any HTML characters.

This is especially important with the libXML errors added in a00068b98,
whose error messages will include literal lines of HTML.