bbify: add [verbatim] tag
This will output the contents inside the tag verbatim in mono-spaced
font. It is needed in particular to avoid expanding other tags when
you want to display them pre-expanded (e.g. documenting new tags in
the changelog).

Note that the contents will be pre-escaped by BBCode, so if the string
is also being escaped locally, any HTML inside the tag will display in
its escaped form, i.e.:

    PB & J [verbatim]PB & J[\verbatim]

will display as

    PB & J <tt>PB &amp; J</tt>