irc: fix a strict SQL error
The `irc_seen.registered_nick` does not have a default value and needs
to be set when a new row is added (the empty string is fine, as it will
be set properly when checking the registration status).

Fixes the following SQL error:

> Field 'registered_nick' doesn't have a default value

We now replace a row when a previously seen user logs in again, so that
we can re-use the same logic as seeing a user for the first time, and
reset some fields to their defaults without needing to specify them
explicitly.