Fix strict type comparison issue with missions
Switching to strict equality operators in f5f23b4f387c resulted in an
"always false" comparison in `checkMissionRequirements` between integer
data and strings inserted into the mission template. As a result, the
mission step would never advance.

To fix this, we now preserve the type when we replace the template if
the placeholder is the entire template (rather than part of a larger
string).