Player: allow saving negative HOF stats
Save all new HOF stats, regardless of the value. Previously, we
required that the value was positive, which meant that if the first
value was negative, it wouldn't be saved.

This is a pretty niche issue, but it impacts your very first trade
(negative Trade:Money:Profit). Once you make a positive profit, you
insert the row, and then rows can be updated regardless of whether
the amount is positive or negative.

The original reason for this insertion condition was likely to avoid
spamming the HOF table with rows of zeros. However, since we return
early from `{increase,decrease,set}HOF` if the value is not changed,
this shouldn't be an issue.