Fix creating a new chess game
Due to the strict SQL changes, not providing a value for `winner_id`
when creating a `chess_game` row caused an error because `winner_id`
did not have a default value. We now set the default value to 0.

Also removed `$endDate` from the `insertNewGame` interface, as it is
not needed. (We only create new games that haven't ended yet.)