Fix creating a new feature request
Due to strict SQL changes, not providing values for `fav`, `yes`, and
`no` when creating a `feature_request` row caused an error because the
columns did not have default values. These tallies were only used for
feature requests that could no longer be voted on, whereas votes for
open requests are tallied dynamically from `account_votes_for_feature`;
we can drop these columns and simply use the latter method for both
open and non-open requests (it is perhaps a bit slower, but it is much
more straightforward).