Fix empty Discord widget in alliance MotD page
In 66045e9c5cf, we changed the condition for displaying the widget from
`!empty` to `isset`, which means that the empty string will result in
displaying a blank widget. We could change the condition to isset and
is not the empty string, but it seems better to canonicalize the empty
string input to null when changing alliance settings.