Player: fix sameAlliance for players not in alliance
The Player::sameAlliance method is expected to return true when the
two players are the same player. This is basically just a shortcut for
avoiding an additional Player::equals check in some places. The issue
is that this was returning false if the player was not in an alliance.

This issue was introduced in c046eea660. It caused solo players to be
hostile to their own forces and to not be able to land on their own
planet.