SmrAlliance: disallow alliance ID of 0
Previously, we would allow creating an `SmrAlliance` instance when the
input `$allianceID` was 0. This was problematic because most of the
class attributes would be unset, and most methods would be unusable.

Now, we will raise an AllianceNotFoundError if the ID is 0, and we
remove the `isNone` method. This will ensure that all instances of the
class are valid and can be used without restriction. We replace the
existing `isNone` calls with checks if the alliance ID is 0.