game_play.php: add note when there are no games
In case the current game ends without one prepared to replace it, we
want some way to direct players to get information about when the next
game will start.

To achieve this, we add a note to the "Play Game" page if there is no
active games and no games to join, with a link to the Discord server to
get more information.