DatabaseRecord: add getClass method
Use the @template syntax to restrict the type of the $class parameter
and return type.

This allows us to ensure that we receive an instance of a specified
class from the database. Previously, we were using getObject directly,
which resulted in a return typehint of "mixed" and no validation of the
returned value at runtime.