DiContainer: replace get/make with getClass/makeClass
Use the @template syntax to narrow the $className parameter and return
type.

This allows us to ensure that we receive an instance of a specified
class from the container. Previously, we allowed using these methods to
get any arbitrary key/value from the container, but now such uses need
to query the underlying container directly (with getContainer->set/get).