Fix bug in ChessGame::isCheckmated
Previously, the condition tested was that if the King couldn't make a
legal move, then it was checkmate. This is wrong. Now, we check if any
piece can make a legal move (since a non-King piece could block the
checkmate).

Also moved the `isCheckmated` method from ChessGame to Board, since it
only depends on the board state (modulo player turn), not the game
itself.