dependabot.yml: use directories and groups (#2041)
The use of `directories` instead of `directory` allows consolidating
updates within the same package ecosystem. We had 4 docker entries with
the same settings, which is now collapsed into a single entry.

The use of `groups` with the "php" pattern should allow all PHP version
updates to occur in a single dependabot PR. This ensures consistency
across the repo.