Deduplicate damage taken display construction (#2043)
The combat templates were all duplicating the exact logic for building
the damage taken display HTML. We refactor this logic into the Template
methods `display{Force}TakenDamage`.

Also fix the messages so that they are properly pluralized.

Note that the kamikaze effect is handled slightly differently. Now, if
there is a kamikaze, it subtracts that from the number of mines that
were destroyed by non-kamikaze. Previously, it only printed the mine
damage if it was greater than kamikaze, but then displayed the total
number of mines lost. I think the new logic is more correct, but from a
practical perspective they are identical, because combat drones do not
currently do both kamikaze and normal damage to mines.