DatabaseCleanupProcessor.php: fix phpstan error
Fix PHPStan warning caused by using `$result` in a way that isn't
understood by PHPStan.

Technically, the existing logic was correct (since Database::write
returns an int and only Database::read does not), but there was no
direct relationship between `$result` and `$method` previously.

Switch to first-class callable syntax and add a helper closure to
process the result into an int in a reliable way. For the Preview, we
now also display the number of rows that would have been deleted,
instead of displaying 0.

Clean up the displayed text to more clearly indicate that the Preview
does not actually do any deletion.