DatabaseRecordTest.php: ignore unused result
PHPStan can't infer that we will be raising a TypeError from the call to
DatabaseRecord::getString because we're relying on the return typehint,
rather than explicitly checking the return type (which seems redundant
to me). Since we are explicitly testing this case with PHPUnit, it is
safe to just ignore this specific PHPStan error.