Fix php 8.4 deprecation in nbbc
The vanilla/nbbc code has a deprecation warning that will be fixed with
the following PR:

https://github.com/vanilla/nbbc/pull/39

While we wait for that to be merged, install from the PR branch to avoid
the deprecation warning.

Added a test of the `bbify` function, since there was previously no test
code exercising nbbc. Also pass the `smrBBCode` function as a first-
class callable when creating BBCode rules.