UniGen: add ability to insert new galaxy
When designing games, it can be helpful to be able to add a new galaxy.
In the past, we would have to create an entirely new game. This feature
is only available for games that are not yet enabled.

It was easiest to create an empty dummy galaxy that could then be
modified using the existing EditGalaxiesProcessor subsequently. I had
considered creating a javascript interface where you specify all desired
changes client-side before any submissions, but that was harder to code
both for the display and the processing. Maybe something for the future.

* Galaxy::getConnectivity now handles 0-size galaxies, since that's the
  default for newly inserted galaxies.

* Galaxy::getStartSector no longer assumes that we have all the galaxy
  IDs starting from 1. This *should* be true, but I ran into this issue
  during development due to an improperly added galaxy, and the new
  implementation handles this case and is also just safer in general.

* Tweaked the warning on the "Edit Galaxies" page so that it makes more
  sense when you can/can't change galaxy sizes (i.e. when the game
  isn't/is enabled respectively).