UniGen: add basic map auto-generation option
Now that we have more flexibility to add/remove galaxies as needed, the
functionality to pre-populate a map using a fairly basic/generic layout
can help save a lot of time.

This can be tweaked as needed, but currently it sets up the following:

* 8 racial galaxies
* 1 racial hub with a planet for the NHL
* 1 large neutral galaxy with most locations
* 4 small planet galaxies

Special care is taken to protect the warps/fed in racial galaxies from
being mined.

Avoided ports in the same galaxies as planets, since this has been
abused in the past, and we'll need a solution to that first.