Gameplay: add No-Drop Zone location
This is a location that can be used to prevent dropping forces without
needing to add a functional location. Usually in the past we would just
spam Bar or Bank locations to achieve the same goal, but that is not
always desirable.

If we ever rework forces in the future, one scenario would be to remove
the force-dropping restriction on locations, but retain it only for NDZ.
This would give more flexibility over where exactly forces are allowed.
To be clear - this is not implemented here.