Location: add static getters for bars/banks/shops
This is a more efficient way to get all locations of a certain type,
which only involves a single SQL query. Compared to getting all of the
locations and then filtering on `isX`, which involves N+1 SQL queries,
where N is the total number of locations.

We do not cache this result, since it is not expected to be called
regularly (compared to, e.g. getLocation).