CheckMap: list unreachable sectors
It's not easy to spot unreachable sectors by eye, so automate it! The
algorithm starts at sector 1 (which we assume exists and is accessible),
and then increments through all sectors in the map. If the sector is
accessible from the most recently accessible sector, then it becomes the
new most recently accessible sector. If not, it's tagged as unreachable.

This is likely the most efficient algorithm on average, since neighbor
sectors are typically the shortest path to each other. (As opposed to,
say, checking all sectors from sector 1, which would do a lot of
redundant work to find many of the shortest paths.