GameDelete: allow deleting not-enabled games
The games list in the UniGen can get very cluttered if we aren't able to
delete games (e.g. broken, abandoned, etc.).

The previous implementation of this feature was for archiving games that
were actually played, and consolidating its data into the history SQL
database. This implementation is different, it's more for cleaning up
the games in the Dev/Beta UniGen. Therefore, we do not allow deleting
games that have been enabled, as a guard against purging games that were
played (given that there is no "archive" functionality).

There are no plans to re-introduce archiving at this time.