Visual: always display menu bar if PageTopic is set
For visual consistency and more structured content presentation, an
empty message bar will be displayed if there is no menu but there is a
PageTopic set.

We could always have the menu bar, but it looks a bit weird on pages
with no menu and no PageTopic (e.g. local map). It's also not ideal in
these cases if the omission is for compactness (e.g. UniGen).