UniGen: view-only option for maps
Generalize EditGalaxy so that the same UniGen interface can be used for
both editing and viewing the galaxy map. This has two main uses:

1. When you want to avoid accidentally changing something (e.g. maybe
   it's a past game or a game in progress), just click the new "View"
   button instead of "Edit".

2. A potential future feature where not all maps are editable by
   everyone. For example, by default maps could be editable only by
   their creators and "super admins".

This was achieved with a `canEdit` boolean property of EditGalaxy. Note
that because we return to the EditGalaxy page from many different pages,
we were passing around the `galaxyID` value so that we could recreate
it. However, now that it has even more properties to track, we don't
want every other page to inherit an additional property. The simple
solution is to pass around the EditGalaxy object to return to, rather
than its constituent properties (as we have done with other pages, with
"forward to" or "target page" Page properties). This also makes it
clearer what we were using `galaxyID` for in pages that had to return to
EditGalaxy -- in most cases it was only for recreating EditGalaxy.

We add a `returnTo` method on the EditGalaxy class to clone the object
without a message, since those should be ephemeral and should not be
displayed when we return from another page.