UniGen: revamp game edit selection
Add table `game_create_status` to track some details about a game during
the creation process (creator, creation date, if anyone can edit it, and
when it was marked ready). These details are now displayed in a table
(rather than a dropdown) on the page where you select which map to edit
in the UniGen. The table can be sorted on any column, and you can also
delete maps from this table.

Setting the edit permissions and marking the map ready are now done in
the page after selecting to edit a map.

Existing non-enabled games will show up with a dummy creation date and
no creator (since that information is lost to time).

The create/edit selection page now also has separate dropdown selections
for viewing past games (no editing allowed, because there's no reason to
ever change a past game) and viewing or editing active games (if you
have permission to edit enabled games).

Also changed the following:

* Added DatabaseRecord::getNullableBoolean, which could be used in, for
  example a LEFT JOIN with a bool column on the right. I only used this
  when I wasn't going to backfill the `game_create_status` table, and
  thus there would be records with no `all_edit` null, but I still left
  it in because it's tested and may be useful in the future.

* Added Game::get{Past,Active}Games static methods. This was implemented
  in the typical way where we pass an existing Record to the ctor, so
  that we're only making 1 query instead of N. Tried to update existing
  `game` table queries to use these when possible.