Increase game_id column size from tinyint to smallint
With the changes to UniGen that facilitate easy creation/deletion of
games, it seems likely that we would hit the `tinyint` limit of 256 in
the near future. With `smallint` the limit is increased to 65536.

This may be a premature optimization, and we should just go with `int`,
but it also seems unlikely we'll ever hit this new limit.