Give mysql service more time to pass healthcheck
In CI, we sometimes fail to start the mysql service in the allotted
time. During the `start_period`, the test will be performed every 5s
(by default) but won't count toward the number of retries. Then we'll
try every 5s 10 times for a total of 30 + 50 = 80s. Timeout is decreased
to 5s because the command should return almost immediately once mysql is
ready for connections.

Note that this was likely failing because the first mysql run requires
additional setup. If we continue to see failures, then there is likely
some other issue at play.