Fix browser translation of input submit elements
When using an `<input type="submit" name=... value=... />` element, the
"value" field is part of both the front-end display (the text displayed
on the submit element) and the back-end processing (the data submitted
with the request). As a result, when browsers automatically translate
the page contents for non-English speakers, they change the value that
is submitted with the request.

To fix this, we convert all named submit inputs to `<button>` elements
instead, since these completely decouple the submitted data from the
displayed text.

Add the `create_submit` function to standardize this process and ensure
that we can modify this behavior universally if needed.

Also add a `create_submit_display` function for cases where there is
only a single submit element in a form. Here we don't need the submit to
supply a name-value pair in the submitted data, and the function input
is used only for display text.