Work around phpstan 2.1.23 issue
Using `array_pop` and `array_unshift` on the same line results in the
following phpstan error (in 2.1.23+):

> Property Smr\AbstractShip::$weapons (array<int, Smr\Combat\Weapon\Weapon>)
> does not accept array<int, Smr\Combat\Weapon\Weapon|null>.

This is a false positive because the array is validated to be non-empty,
and so the `array_pop` value will never be null.

For some reason, doing the `array_pop` in a separate expression first
avoids this error.

See https://github.com/phpstan/phpstan/issues/13510.