Fix discord bot in server channels
A few years ago, Discord made server channel messages a privileged
intend, meaning we needed to request that permission explicitly
(which I did, but note this needs to be approved by Discord if the bot
is in >100 servers). However, we also needed to request that permission
in DiscordPHP when running the bot (which I did not do).

With the MESSAGE_CONTENT intent specified, the bot will now respond to
messages in server channels.

See https://dis.gd/mcfaq.