Fix exploit in council veto (#2251)
When processing a veto, we were deleting the current (race1, race2)
and (race2, race1) votes. While this is the correct thing to do for
peace votes, it is incorrect for war votes, because the two races may
have separate war votes in progress.

This resulted in an exploit where the president of race A could start
and then immediately veto a war vote with race B to cancel race B's
war vote against race A. Now, we only delete both votes if it's a peace
vote. Thanks to JettJackson for reporting this long-standing issue!

The one remaining pathology is that the president of race A can infer if
there is an active war vote against them by trying to start a peace vote
with race B. We currently can't allow both an active peace vote and war
vote between two races due to the uniqueness of the (race1, race2) pair
in the database (regardless of the type). This will require changes to
the `race_has_voting` and `player_votes_pact` table structure, probably
including `type` in the primary key, along with major PHP logic changes.

Some shared code (counting and deleting votes) was refactored out into
common functions in `CouncilVoting` to avoid code duplication.